test(textarea): migrate color
