+++ core/src/components/textarea/test/color/textarea.e2e.ts-snapshots/textarea-no-fill-color-end-md-ltr-Mobile-Chrome-linux.png
