+++ core/src/components/textarea/test/color/textarea.e2e.ts-snapshots/textarea-no-fill-color-fixed-md-ltr-Mobile-Safari-linux.png
