+++ core/src/components/textarea/test/color/textarea.e2e.ts-snapshots/textarea-outline-color-end-md-ltr-Mobile-Firefox-linux.png
