+++ core/src/components/textarea/test/color/textarea.e2e.ts-snapshots/textarea-outline-color-stacked-md-ltr-Mobile-Chrome-linux.png
