+++ core/src/components/textarea/test/color/textarea.e2e.ts-snapshots/textarea-solid-color-fixed-md-ltr-Mobile-Firefox-linux.png
