+++ core/src/components/textarea/test/color/textarea.e2e.ts-snapshots/textarea-solid-color-floating-md-ltr-Mobile-Safari-linux.png
