+++ core/src/components/textarea/test/color/textarea.e2e.ts-snapshots/textarea-solid-color-stacked-md-ltr-Mobile-Firefox-linux.png
