+++ core/src/components/textarea/test/color/textarea.e2e.ts-snapshots/textarea-solid-color-start-md-ltr-Mobile-Safari-linux.png
